KSY-알림장_학부모_전체조회_로직_구현

+++ src/main/java/com/aico/aibayo/entity/AcceptLogEntity.java
@@ -17,7 +17,7 @@ public class AcceptLogEntity {
     @Id
     @Column(name = "accept_no")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int acceptNo;
+    private Long acceptNo;
     @Column(name = "code_verify")
     private int codeVerify;
     @Column(name = "accept_type")
